This PR saves some allocations on `MSTORE`, which seems to be the second most common opcode. 

```
name        old time/op    new time/op    delta
OpMstore-2    61.7ns ± 4%    30.9ns ± 1%   -49.88%  (p=0.000 n=10+9)

name        old alloc/op   new alloc/op   delta
OpMstore-2     32.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)

name        old allocs/op  new allocs/op  delta
OpMstore-2      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
```
 It also contains a minor optimization to `SLOAD`, to reuse bigints and save a malloc/copy operation.