@karalabe 2018-06-14
@fjl Could you open a PR with the fix?